+++ src/main/kotlin/dev/thiagosouto/plugins/bom/models.kt
@@ -1,12 +1,18 @@
 package dev.thiagosouto.plugins.bom
 
-data class BomInfo(val projectAttributes: List<Tag>, val tags: List<Tag>, val dependencies: List<Dependency>)
-data class Tag(val name: String, val value: String)
-data class Dependency(
+internal data class BomInfo(val projectAttributes: List<Tag>, val tags: List<Tag>, val dependencies: List<Dependency>)
+internal class SimpleTag(override val name: String, val value: String) : Tag
+
+internal class RootTag(override val name: String, val tags: List<Tag>) : Tag
+
+internal data class Dependency(
     val artifactId: String,
     val groupId: String,
     val version: String,
     val exclusions: List<Exclusion> = emptyList()
 )
 
-data class Exclusion(val artifactId: String, val groupId: String)
+internal data class Exclusion(val artifactId: String, val groupId: String)
+interface Tag {
+    val name: String
+}