+++ src/main/kotlin/dev/thiagosouto/plugins/bom/pom/BOMDocumentCreator.kt
@@ -3,6 +3,8 @@ package dev.thiagosouto.plugins.bom.pom
 import dev.thiagosouto.plugins.bom.BomInfo
 import dev.thiagosouto.plugins.bom.Dependency
 import dev.thiagosouto.plugins.bom.Exclusion
+import dev.thiagosouto.plugins.bom.RootTag
+import dev.thiagosouto.plugins.bom.SimpleTag
 import dev.thiagosouto.plugins.bom.Tag
 import org.w3c.dom.Document
 import org.w3c.dom.Element
@@ -10,7 +12,7 @@ import org.w3c.dom.Node
 import javax.xml.parsers.DocumentBuilder
 import javax.xml.parsers.DocumentBuilderFactory
 
-class BOMDocumentCreator {
+internal class BOMDocumentCreator {
 
     fun create(bomInfo: BomInfo): Node {
         val docBuilder: DocumentBuilder = DocumentBuilderFactory.newInstance().newDocumentBuilder()
@@ -31,13 +33,22 @@ class BOMDocumentCreator {
 
     private fun createMavenAttributes(rootElement: Element, projectAttributes: List<Tag>) {
         for (attribute in projectAttributes) {
-            rootElement.setAttribute(attribute.name, attribute.value)
+            if (attribute is SimpleTag) {
+                rootElement.setAttribute(attribute.name, attribute.value)
+            }
         }
     }
 
     private fun appendPomInfo(document: Document, rootElement: Element, tags: List<Tag>) {
         for (tag in tags) {
-            rootElement.appendChild(createTagValueElement(document, tag.name, tag.value))
+            when (tag) {
+                is SimpleTag -> rootElement.appendChild(createTagValueElement(document, tag.name, tag.value))
+                is RootTag -> {
+                    val rootTag = document.createElement(tag.name)
+                    appendPomInfo(document, rootTag, tag.tags)
+                    rootElement.appendChild(rootTag)
+                }
+            }
         }
     }